# specs/001-chronicleai-publication-platform/contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: ChronicleAI API
  version: 0.1.0
  description: Contract for ChronicleAI public alerts, premium intelligence, KeeperHub ingestion, payments, and public agent activity data.
servers:
  - url: https://api.chronicleai.example
    description: Production API
  - url: http://localhost:4000
    description: Local API
paths:
  /health:
    get:
      summary: Check API health
      responses:
        "200":
          description: API is healthy
  /alerts:
    get:
      summary: List public alerts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Public alerts
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicAlert"
  /digests/latest:
    get:
      summary: Get latest public digest
      responses:
        "200":
          description: Latest digest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyDigest"
        "404":
          description: No digest available
  /subscribers:
    post:
      summary: Free email opt-in for alerts (and optional digests)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                receivesDigests:
                  type: boolean
                receivesAlerts:
                  type: boolean
                payerReference:
                  type: string
                source:
                  type: string
                  enum: [web, api, premium, import]
      responses:
        "200":
          description: Existing subscriber updated
        "201":
          description: Subscriber created
  /subscribers/newsletter/subscribe:
    post:
      summary: Start a recurring x402 monthly newsletter agreement (issues payment challenge)
      description: |
        Creates or re-opens a newsletter subscription agreement and returns an x402
        TransferWithAuthorization challenge for the first billing period.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                payerReference:
                  type: string
                  description: EVM wallet that will sign the x402 authorization
                referralAddress:
                  type: string
                  description: Optional affiliate wallet for revenue attribution
      responses:
        "201":
          description: Newsletter x402 challenge issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterChallenge"
        "400":
          description: Invalid request or already entitled for the current period
  /subscribers/newsletter/renew:
    post:
      summary: Issue an x402 renewal challenge for the next billing period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                payerWallet:
                  type: string
      responses:
        "201":
          description: Renewal challenge issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterChallenge"
  /subscribers/newsletter/settlements:
    post:
      summary: Settle an x402 newsletter challenge and activate or extend the billing period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeReference, settlementReference]
              properties:
                challengeReference:
                  type: string
                settlementReference:
                  type: string
      responses:
        "200":
          description: Settlement result and subscription state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSettlement"
  /subscribers/newsletter/status:
    get:
      summary: Get recurring newsletter subscription status and entitlement
      parameters:
        - name: email
          in: query
          schema:
            type: string
            format: email
        - name: payerWallet
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscription"
        "404":
          description: Subscription not found
  /subscribers/newsletter/cancel:
    post:
      summary: Cancel a recurring newsletter subscription (at period end by default)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                payerWallet:
                  type: string
                atPeriodEnd:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscription"
  /premium/items:
    get:
      summary: List public metadata for premium intelligence
      responses:
        "200":
          description: Premium item teasers
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/PremiumItemTeaser"
  /premium/items/{id}:
    get:
      summary: Get premium intelligence content after payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paid premium content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremiumItem"
        "402":
          description: Payment required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentChallenge"
        "403":
          description: Payment invalid or expired
        "404":
          description: Premium item not found
  /payments/challenges:
    post:
      summary: Create a payment challenge for premium content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [premiumItemId, paymentRoute]
              properties:
                premiumItemId:
                  type: string
                paymentRoute:
                  type: string
                  enum: [x402, mpp]
                payerReference:
                  type: string
      responses:
        "201":
          description: Payment challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentChallenge"
  /payments/settlements:
    post:
      summary: Submit or receive payment settlement evidence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeReference, paymentRoute, settlementReference]
              properties:
                challengeReference:
                  type: string
                paymentRoute:
                  type: string
                  enum: [x402, mpp]
                settlementReference:
                  type: string
                amountSettled:
                  type: number
                currency:
                  type: string
      responses:
        "200":
          description: Settlement evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRecord"
  /keeperhub/events:
    post:
      summary: Ingest monitored on-chain event from KeeperHub
      security:
        - KeeperHubWebhook: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeeperHubEvent"
      responses:
        "202":
          description: Event accepted for processing
        "400":
          description: Invalid event payload
        "401":
          description: Invalid webhook authentication
        "409":
          description: Duplicate event already received
  /keeperhub/digests/run:
    post:
      summary: Trigger daily digest generation from KeeperHub scheduled workflow
      security:
        - KeeperHubWebhook: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [periodStart, periodEnd]
              properties:
                periodStart:
                  type: string
                  format: date-time
                periodEnd:
                  type: string
                  format: date-time
      responses:
        "202":
          description: Digest generation accepted
  /keeperhub/treasury/check:
    post:
      summary: Trigger treasury sustainability check from KeeperHub scheduled workflow
      security:
        - KeeperHubWebhook: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capturedAt, availableBalance, currency, safetyBuffer]
              properties:
                capturedAt:
                  type: string
                  format: date-time
                availableBalance:
                  type: number
                currency:
                  type: string
                safetyBuffer:
                  type: number
      responses:
        "202":
          description: Treasury snapshot accepted
        "400":
          description: Snapshot invalid
        "401":
          description: Invalid webhook authentication
  /keeperhub/revenue/route:
    post:
      summary: Trigger weekly autonomous revenue routing from KeeperHub scheduled workflow
      security:
        - KeeperHubWebhook: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [periodStart, periodEnd]
              properties:
                periodStart:
                  type: string
                  format: date-time
                periodEnd:
                  type: string
                  format: date-time
      responses:
        "202":
          description: Revenue routing accepted
        "400":
          description: Invalid payload or period
        "401":
          description: Invalid webhook authentication
        "409":
          description: Payouts for this period have already been routed
  /activity:
    get:
      summary: Get public agent activity dashboard data
      responses:
        "200":
          description: Public agent activity data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentActivity"
components:
  securitySchemes:
    KeeperHubWebhook:
      type: apiKey
      in: header
      name: X-ChronicleAI-Signature
  schemas:
    PublicAlert:
      type: object
      required: [id, title, summary, sourceReferences, deliveryStatus, publishedAt]
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        sourceReferences:
          type: array
          items:
            type: string
        deliveryStatus:
          type: string
          enum: [draft, queued, published, partial_failure, failed]
        publishedAt:
          type: string
          format: date-time
        confidence:
          type: string
          enum: [high, medium, low]
        generationProvider:
          type: string
          enum: [gemini, openai, groq]
        generationAttempts:
          type: array
          items:
            $ref: "#/components/schemas/LLMGenerationAttempt"
        registryTxHash:
          type: string
        sourceEventHash:
          type: string
        contentUri:
          type: string
    DailyDigest:
      type: object
      required: [id, reportDate, title, summary, highlights, publicationStatus]
      properties:
        id:
          type: string
        reportDate:
          type: string
          format: date
        title:
          type: string
        summary:
          type: string
        highlights:
          type: array
          items:
            type: string
        publicationStatus:
          type: string
          enum: [draft, queued, published, partial_failure, failed]
        registryTxHash:
          type: string
        sourceEventRoot:
          type: string
        contentUri:
          type: string
    PremiumItemTeaser:
      type: object
      required: [id, title, summaryPublic, priceAmount, priceCurrency, paymentRoutes]
      properties:
        id:
          type: string
        title:
          type: string
        summaryPublic:
          type: string
        priceAmount:
          type: number
        priceCurrency:
          type: string
        paymentRoutes:
          type: array
          items:
            type: string
            enum: [x402, mpp]
    PremiumItem:
      allOf:
        - $ref: "#/components/schemas/PremiumItemTeaser"
        - type: object
          required: [contentPrivate, sourceReferences]
          properties:
            contentPrivate:
              type: object
            sourceReferences:
              type: array
              items:
                type: string
    PaymentChallenge:
      type: object
      required: [challengeReference, premiumItemId, paymentRoute, amountRequested, currency, expiresAt]
      properties:
        challengeReference:
          type: string
        premiumItemId:
          type: string
        paymentRoute:
          type: string
          enum: [x402, mpp]
        amountRequested:
          type: number
        currency:
          type: string
        expiresAt:
          type: string
          format: date-time
    PaymentRecord:
      type: object
      required: [id, premiumItemId, paymentRoute, status]
      properties:
        id:
          type: string
        premiumItemId:
          type: string
        paymentRoute:
          type: string
          enum: [x402, mpp]
        status:
          type: string
          enum: [challenge_issued, pending, settled, underpaid, expired, failed]
        settlementReference:
          type: string
    KeeperHubEvent:
      type: object
      required: [sourceEventId, eventType, chainId, capturedAt, rawPayload]
      properties:
        sourceEventId:
          type: string
        eventType:
          type: string
          enum: [large_swap, liquidation, gas_spike, volume_anomaly, contract_deployment]
        chainId:
          type: integer
        protocol:
          type: string
        transactionHash:
          type: string
        magnitude:
          type: object
          properties:
            value:
              type: number
            unit:
              type: string
        capturedAt:
          type: string
          format: date-time
        rawPayload:
          type: object
    LLMGenerationAttempt:
      type: object
      required: [provider, attemptOrder, status, latencyMs]
      properties:
        provider:
          type: string
          enum: [gemini, openai, groq]
        attemptOrder:
          type: integer
          minimum: 1
          maximum: 3
        status:
          type: string
          enum: [succeeded, failed, invalid_response]
        latencyMs:
          type: integer
          minimum: 0
        failureReason:
          type: string
    AgentActivity:
      type: object
      required: [alerts, digests, payments, treasury, executionLogs, sponsoredWatches, payouts]
      properties:
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/PublicAlert"
        digests:
          type: array
          items:
            $ref: "#/components/schemas/DailyDigest"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentRecord"
        treasury:
          type: object
          required: [availableBalance, safetyBuffer, status]
          properties:
            availableBalance:
              type: number
            safetyBuffer:
              type: number
            status:
              type: string
              enum: [healthy, warning, critical]
        executionLogs:
          type: array
          items:
            type: object
        sponsoredWatches:
          type: array
          items:
            $ref: "#/components/schemas/SponsoredWatch"
        payouts:
          type: array
          items:
            $ref: "#/components/schemas/RevenuePayout"
    SponsoredWatch:
      type: object
      required: [id, targetContract, watchSpecHash, startsAt, endsAt, createTxHash, status]
      properties:
        id:
          type: string
        targetContract:
          type: string
        watchSpecHash:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        createTxHash:
          type: string
        reportTxHash:
          type: string
        reportContentHash:
          type: string
        contentUri:
          type: string
        status:
          type: string
          enum: [accepted, monitoring, completed, failed]
    RevenuePayout:
      type: object
      required: [id, payoutPeriodHash, recipient, amount, reasonHash, payoutTxHash, registryTxHash, status]
      properties:
        id:
          type: string
        payoutPeriodHash:
          type: string
        recipient:
          type: string
        amount:
          type: number
        reasonHash:
          type: string
        payoutTxHash:
          type: string
        registryTxHash:
          type: string
        status:
          type: string
          enum: [pending, transferred, failed]
    NewsletterSubscription:
      type: object
      required:
        - id
        - email
        - status
        - amountPerPeriod
        - currency
        - billingPeriodDays
        - periodsPaid
        - cancelAtPeriodEnd
        - entitled
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [pending, active, past_due, cancelled, expired]
        amountPerPeriod:
          type: number
        currency:
          type: string
        billingPeriodDays:
          type: integer
        currentPeriodStart:
          type: string
          format: date-time
          nullable: true
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
        nextRenewalAt:
          type: string
          format: date-time
          nullable: true
        periodsPaid:
          type: integer
        payerWallet:
          type: string
          nullable: true
        referralAddress:
          type: string
          nullable: true
        cancelAtPeriodEnd:
          type: boolean
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        lastSettledAt:
          type: string
          format: date-time
          nullable: true
        entitled:
          type: boolean
          description: True when the subscriber may receive premium digests now
    NewsletterChallenge:
      type: object
      required:
        - subscriptionId
        - email
        - status
        - challengeReference
        - paymentRoute
        - amountRequested
        - currency
        - expiresAt
        - challengeData
        - paymentRecordId
        - billingPeriodDays
        - agreementType
      properties:
        subscriptionId:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [pending, active, past_due, cancelled, expired]
        challengeReference:
          type: string
        paymentRoute:
          type: string
          enum: [x402]
        amountRequested:
          type: number
        currency:
          type: string
        expiresAt:
          type: string
          format: date-time
        challengeData:
          type: object
          additionalProperties: true
        paymentRecordId:
          type: string
        billingPeriodDays:
          type: integer
        agreementType:
          type: string
          enum: [recurring_newsletter]
        periodKind:
          type: string
          enum: [initial, renewal]
    NewsletterSettlement:
      type: object
      required: [settled, subscription, paymentRecordId, verification]
      properties:
        settled:
          type: boolean
        subscription:
          $ref: "#/components/schemas/NewsletterSubscription"
        paymentRecordId:
          type: string
        verification:
          type: object
          required: [amountSettled, currency, settlementReference]
          properties:
            amountSettled:
              type: number
            currency:
              type: string
            settlementReference:
              type: string
            payerReference:
              type: string
            errorMessage:
              type: string